Name copied animations uniquely
This fixes a bug that caused deleted keyframes in copied sequences to not be removed from the InteractiveTree. Since the animation in the original sequence and in the copy had the exact same name, the algorithm to patch the tree could never find the animation in the copied sequence.

Also adds a regression test. Empty for now but should be added in the future.