Scale fonts in Interactive based on a per-morph static ratio
The previous scaling mechanism was easy to throw of. In addition to the usage of scale to zoom in the interactive, this change should make it more stable.
We calculate the fontSize - Interactive height ratio upon adding the morph to the interactive and keep that consistent over the lifetime of the morph in the interactive. Thus, scaling a text down to fontSize 0 and up again will not "reset" the ratio anymore and the fontSize in the end will be consistent with the one before scaling the text down.