Show border between InteractiveHolder and Graph in emtpy editor
This is not the cleanest solution, as it utilizes the border of the InteractiveHolder, whereas one would like to use the graph.
However, this is currently not possible, due to the way the graph is layoutet.
I opted for this solution as tackling the problem yields improved readability of the UI. A cleaner solution should be revisited in the future.