Rebind active halos upon hovering into editor
This would lead to the halo shown when clicking in halo mode being slimmer than the default halo. The default halo can always be brought back with Ctrl+Click.
However, halo mode does currently not work correctly due to the editor being in a window and additionally the event delegation in the ScrollHolder.