Switch to Halo mode when forcefully switching TopBarMode
When the editor changes the topbar mode (e.g., when hovering into an interactive in global view but with a morph-creation mode selected in the topbar), we now switch to halo mode instead of hand mode.
This is because with the tighter integration into the styling palette, the halo mode becomes essential to edit morphs in the interactive.