Enable most features of new animation creation workflow
All references to old control widgets of inspector should now be removed.
The task of the widgets is now solely to animate content.
A widget can either be created because an animation in a selected morph already exists, via the dropdown or because a property of the targetmorph was manipulated. In this case, the widget is marked.